+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changed
 - Use a single database instead of multiple
 - Fixed subtitle loading
+- Show a loading progress indicator until search is finished on the search screen
 
 ## [0.14.1] - 2022-05-04
 ### Changed